chore: move files to the top level of the repository
Extract files from the place they used to be in my dotfiles.

I initially started working on this plugin as part of my dotfiles. When
it grew large enough, I extracted it to a separate repository by
cherry-picking commits. This meant the file paths were as in my
dotfiles. This is not intended when those files are in a standalone
repository. They should be at the top level.