refactor(pcomb): split symbols into individual files
Follow the module structure of nom [0] and split parsers into separate
files. This way they are easier to work on, since it is not a single
huge file with all of them.

All of the submodules are merged in lua/pcomb/init.lua.
Unfortunately, since this is dynamic, Lua-LS cannot retrieve type
information for `require("pcomb")`.

[0]: https://docs.rs/nom/latest/nom/index.html