refactor(natdat): extract time parsers and improve output structure
Use Time24H and TimeAMPM classes to represent the two formats of time
that natdat understands.

Each has a `format` method to convert it into a string and a
`to_24h_time` method to use later when computing the timestamp.

This temporarily breaks the library, since no code uses the new parsers,
and the old ones are removed.